router.py and host.py added (comments from Laksh and Yubo).
Separated Device into subclasses. Modified typos and errors.
Ready for review.
