+++ README.md
@@ -28,15 +28,15 @@ be installed.
   http://unix.stackexchange.com/questions/164517/nonetype-object-has-no-attribute-decompressobj-while-installing-bootstrap-se
 ```
 
-- setuptools
+- setuptools:
 ```
 https://pypi.python.org/pypi/setuptools#unix-wget
 ```
 
 - TODO(team): More to come.
 
-** Note that zlib1g-dev needs to be install prior to Python installation. If
-the order is reversed, the Python structure might need a re-"make install".
+** Note that zlib1g-dev needs to be installed prior to Python installation.
+If the order is reversed, the Python structure might need a re-"make install".
 
 
 Usage
