+++ network_simulator/common.py
@@ -1,14 +1,14 @@
 """This file contains definition of constants.
 """
 
-ACK_PACKET_SIZE_BITS = 512
 """Size of :class:`.AckPacket` is 512 bits."""
+ACK_PACKET_SIZE_BITS = 512
 
-DATA_PACKET_SIZE_BITS = 8192
 """Size of :class:`.DataPacket` is 8192 bits."""
+DATA_PACKET_SIZE_BITS = 8192
 
-ROUTER_PACKET_SIZE_BITS = 512
 """Size of :class:`.RouterPacket` is 512 bits."""
+ROUTER_PACKET_SIZE_BITS = 512
 
 # TODO(team): In our doc we have, "hard code a constant
 # to use new routing table after some time delay" in our
